+++ src/main/java/com/goodperson/layered/controller/GuestbookController.java
@@ -10,6 +10,8 @@ import com.goodperson.layered.dto.Guestbook;
 import com.goodperson.layered.service.GuestbookService;
 import com.goodperson.layered.service.StateService;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -29,6 +31,8 @@ public class GuestbookController {
     @Autowired
     private StateService stateService;
 
+    private Logger logger = LoggerFactory.getLogger(this.getClass());
+
     @GetMapping(path = "list")
     public String list(@RequestParam(required = false, defaultValue = "0") int start, Model model,
             @CookieValue(name = "firstVisit", defaultValue = "true", required = true) String firstVisit,
@@ -38,9 +42,8 @@ public class GuestbookController {
         int guestbookCount = guestbookService.getCount();
         List<Integer> pageStartList = guestbookService.getPageStartList(guestbookCount);
         boolean isFirstVisit = stateService.getUserIsFirstVisitFromCookieAndResendUpdatedCookie(firstVisit, response);
-        System.out.println("-----------user-agent-------------");
-        System.out.println(headerInfo.get("user-agent"));
-        System.out.println("---------------------------------");
+        logger.debug("\n-----------user-agent-------------\n{}\n---------------------------------",
+                headerInfo.get("user-agent"));
         model.addAttribute("list", list);
         model.addAttribute("count", guestbookCount);
         model.addAttribute("pageStartList", pageStartList);
@@ -56,21 +59,26 @@ public class GuestbookController {
     // 방명록 작성 함수
     @PostMapping("/write")
     public String write(@ModelAttribute Guestbook guestbook, HttpServletRequest request) {
+        if (guestbook == null)
+            return "redirect:list";
         String clientIp = request.getRemoteAddr();
-        System.out.println("clientIp: " + clientIp);
+        logger.debug("clientIp: {}", clientIp);
         guestbookService.addGuestbook(guestbook, clientIp);
         return "redirect:list";
     }
 
     @GetMapping("/delete")
-    public String delete(@RequestParam(required = true) long id, @SessionAttribute("isAdmin") String isAdmin,
+    public String delete(@RequestParam(required = false) Long id, @SessionAttribute(required = false) String isAdmin,
             HttpServletRequest request, RedirectAttributes redirectAttr) {
+        if (id == null) {
+            return "redirect:list";
+        }
         if (isAdmin == null || !"true".equals(isAdmin)) {
             redirectAttr.addFlashAttribute("errorMessage", "로그인을 하지 않았습니다.");
             return "redirect:loginform";
         }
         String clientIp = request.getRemoteAddr();
-        System.out.println("clientIp: " + clientIp);
+        logger.debug("clientIp: {}", clientIp);
         guestbookService.deleteGuestbook(id, clientIp);
         return "redirect:list";
     }
